user controller profile method created
