+++ watches/app/Http/Controllers/WatchController.php
@@ -118,30 +118,7 @@ class WatchController extends Controller
         
         $title = 'contact';
         return view('contact');
-    }
-
-    /**
-     * Display profile page with previous orders
-     *
-     * @return view 'profile'
-     */
-    public function profile()
-    {
-        $title = 'Profile';
-        //get authenticated user id
-        $id = Auth::id();
-        // test if user is authenticated
-        if(empty($id)){
-            return view('/auth/login'); // redirect to login if not
-        }
-        // get user's data
-        $user = User::find($id);
-        // get all orders associated with the user
-        $orders = Order::all()->where('user_id', $id);
-
-        // return view with user's profile
-        return view('/profile', compact('title','user','orders'));
-    }    
+    } 
 
 
 }
